% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{opm_pal}
\alias{opm_pal}
\title{OPM palettes with ramped colours}
\usage{
opm_pal(palette = "primary", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'opm_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
OPM palettes with ramped colours
}
\examples{
library(scales)
show_col(opm_pal()(10))

filled.contour(volcano,color.palette = opm_pal(), asp=1)

}
