% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{opm_palettes}
\alias{opm_palettes}
\title{OPM Color Palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
opm_palettes
}
\description{
A collection of color palettes for the US Office of Personnel Management. The palettes include:
primary 
secondary 
opm_blue_3 
opm_blue_5
opm_light_blue_3
opm_light_blue_5
opm_red_3
opm_red_5
opm_gray_4
}
\examples{

# Make an x-y plot using the Primary palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
                y = rnorm(100, 0, 20),
               cl = sample(letters[1:8], 100, replace=T))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) + scale_color_mncolors(palette="primary") +
  theme_bw() + theme(aspect.ratio=1)

# Make a histogram using the Secondary palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_mncolors(palette="secondary")

}
\keyword{datasets}
