% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_opm}
\alias{scale_fill_opm}
\title{#' Setup fill palette for ggplot2}
\usage{
scale_fill_opm(
  palette = "primary",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'opm_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
#' Setup fill palette for ggplot2
}
